+++ lib/src/regions/regions.dart
@@ -2,18 +2,23 @@ import 'dart:math';
 import 'package:bishop/bishop.dart';
 
 part 'area.dart';
+part 'built_region.dart';
 part 'rect_region.dart';
 part 'intersect_region.dart';
 part 'region_effect.dart';
+part 'subtract_region.dart';
 part 'union_region.dart';
+part 'xor_region.dart';
 
 abstract class Region {
+  const Region();
   bool contains(int file, int rank);
   Iterable<int> squares(BoardSize size);
   Region translate(int x, int y);
 }
 
 abstract class BoardRegion extends Region {
+  const BoardRegion();
   factory BoardRegion.fromJson(Map<String, dynamic> json) {
     String? type = json['type'];
     if (type != null) {
@@ -25,7 +30,9 @@ abstract class BoardRegion extends Region {
   static const _builders = {
     'rect': RectRegion.fromJson,
     'union': UnionRegion.fromJson,
-    'intersect': IntersectRegion.fromJson
+    'intersect': IntersectRegion.fromJson,
+    'sub': SubtractRegion.fromJson,
+    'xor': XorRegion.fromJson,
   };
 
   Map<String, dynamic> toJson();
@@ -41,6 +48,16 @@ abstract class BoardRegion extends Region {
 
   factory BoardRegion.rank(int rank) => RectRegion.rank(rank);
   factory BoardRegion.file(int file) => RectRegion.file(file);
+
+  BuiltRegion build(BoardSize size) =>
+      BuiltRegion(squares(size).toList(), size);
+
+  UnionRegion operator +(BoardRegion other) => UnionRegion([
+        this,
+        if (other is UnionRegion) ...other.regions else other,
+      ]);
+  SubtractRegion operator -(BoardRegion other) => SubtractRegion(this, other);
+  XorRegion operator ^(BoardRegion other) => XorRegion(this, other);
 }
 
 class BoardRegionAdapter extends BishopTypeAdapter<BoardRegion> {